requirements.txt
streamlit==1.16.0
pandas==1.3.5
numpy==1.24.1
Pillow==9.3.0
plotly==5.11.0